add workflow sub-module
